+++ R/discretize.R
@@ -56,7 +56,7 @@ discretize <- function(x, method = "frequency", breaks = 3,
   )
   
   if(any(duplicated(breaks))){ 
-    stop("The calculated breaks are: ", paste(breaks, collapse = ", "), "\n  Some breaks are not unique. Change the number of breaks or consider using method 'fixed'.")
+    stop("The calculated breaks are: ", paste(breaks, collapse = ", "), "\n  Some breaks are not unique. Look at the distribution of the data (e.g., histogram) to determine appropriate breaks and use the discretization method 'fixed'.")
   }
 
   
@@ -99,7 +99,7 @@ discretizeDF <- function(df, methods = NULL, default = NULL) {
 
 .rediscretizeDF <- function(data, newdata) {
   
-  if(!all(colnames(data) == colnames(newdata))) stop("columns in data and newdata do not conform!")
+  if(!all(colnames(data) == colnames(newdata))) stop("column names in the new data are not the same as in the discretized data.")
   
   cps <- lapply(data, FUN = function(x) {
     breaks <- attr(x, "discretized:breaks")